+++ src/main/java/polytech/ig5/bivouapp/repositories/DisponibilityRepository.java
@@ -3,5 +3,5 @@ package polytech.ig5.bivouapp.repositories;
 import polytech.ig5.bivouapp.models.Disponibility;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DisponibilityRepository extends JpaRepository<Disponibility,Long> {
+public interface DisponibilityRepository extends JpaRepository<Disponibility,Integer> {
 }
